+++ PCIT_core/internal/llvm_interface/src/Function.cpp
@@ -61,5 +61,9 @@ namespace pcit::llvmint{
 
 
 
+	Function::operator Value() const { return Value(static_cast<llvm::Value*>(this->native())); }
+
+
+
 	
 }